♻️ Refactor: 이제 scrap데이터를  저장할  때 유저의 ID를 추가로 저장합니다.
