+++ src/main/java/com/likelion/backendplus4/talkpick/backend/news/info/presentation/controller/NewsInfoDetailProviderController.java
@@ -15,12 +15,13 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.likelion.backendplus4.talkpick.backend.common.annotation.security.LoginUser;
 import com.likelion.backendplus4.talkpick.backend.common.response.ApiResponse;
 import com.likelion.backendplus4.talkpick.backend.news.info.application.dto.NewsInfoDetailResponse;
 import com.likelion.backendplus4.talkpick.backend.news.info.application.port.in.NewsInfoDetailProviderUseCase;
 import com.likelion.backendplus4.talkpick.backend.news.info.application.port.in.NewsViewCountIncreaseUseCase;
 import com.likelion.backendplus4.talkpick.backend.news.info.domain.model.NewsInfoComplete;
-import com.likelion.backendplus4.talkpick.backend.news.info.presentation.dto.ScrapRequest;
+import com.likelion.backendplus4.talkpick.backend.news.info.presentation.controller.dto.request.ScrapRequest;
 import com.likelion.backendplus4.talkpick.backend.news.info.presentation.validator.NewsIdConstraint;
 
 import jakarta.servlet.http.HttpServletRequest;
@@ -54,7 +55,7 @@ public class NewsInfoDetailProviderController {
 	 */
 	@GetMapping("/{id}")
 	public ResponseEntity<ApiResponse<NewsInfoDetailResponse>> getNewsInfoDetailsByArticleId(
-		@PathVariable @NewsIdConstraint String id,  // @NewsIdConstraint 추가
+		@PathVariable @NewsIdConstraint String id,
 		HttpServletRequest request) {
 
 		String ipAddress = request.getRemoteAddr();
@@ -70,9 +71,10 @@ public class NewsInfoDetailProviderController {
 	@PostMapping("/{newsId}/scrap")
 	public ResponseEntity<ApiResponse<Void>> saveScrap(
 		@NotBlank(message = "newsId는 필수입니다.") @PathVariable String newsId,
+		@LoginUser Long loginUser,
 		@Valid @RequestBody ScrapRequest scrapRequest) {
 
-		newsInfoDetailProviderUseCase.saveScrap(toCommand(newsId, scrapRequest));
+		newsInfoDetailProviderUseCase.saveScrap(toCommand(newsId, loginUser, scrapRequest));
 
 		return success();
 	}
