+++ src/main/java/com/likelion/backendplus4/talkpick/backend/news/info/presentation/controller/dto/response/ScrapInfoResponse.java
@@ -1,4 +1,4 @@
-package com.likelion.backendplus4.talkpick.backend.news.info.presentation.dto;
+package com.likelion.backendplus4.talkpick.backend.news.info.presentation.controller.dto.response;
 
 import java.time.LocalDateTime;
