+++ src/main/java/com/likelion/backendplus4/talkpick/backend/news/info/presentation/mapper/NewsInfoResponseMapper.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.likelion.backendplus4.talkpick.backend.common.response.SliceResponse;
 import com.likelion.backendplus4.talkpick.backend.news.info.domain.model.NewsInfo;
 import com.likelion.backendplus4.talkpick.backend.news.info.infrastructure.jpa.adapter.dto.SliceResult;
-import com.likelion.backendplus4.talkpick.backend.news.info.presentation.dto.NewsInfoResponse;
+import com.likelion.backendplus4.talkpick.backend.news.info.presentation.controller.dto.response.NewsInfoResponse;
 
 public class NewsInfoResponseMapper {
 	public static SliceResponse<NewsInfoResponse> toSliceResponse(SliceResult<NewsInfo> domains) {
